backend/requirements.txt
celery==5.6.2
Django==5.2.1
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
openai==1.82.1
python-dotenv==1.1.0
redis==7.1.0
reportlab==4.4.9
